fix and create
fix application.properties for database, create methode edit for Product
