+++ src/main/java/com/telran/store/mapper/ProductMapper.java
@@ -3,12 +3,18 @@ package com.telran.store.mapper;
 import com.telran.store.dto.ProductCreateDto;
 import com.telran.store.dto.ProductResponseDto;
 import com.telran.store.entity.Product;
+import org.mapstruct.BeanMapping;
 import org.mapstruct.Mapper;
+import org.mapstruct.MappingTarget;
+import org.mapstruct.NullValuePropertyMappingStrategy;
 
 import java.util.List;
 
 @Mapper(componentModel = "spring", uses = {CategoryMapper.class})
 public interface ProductMapper {
+
+    @BeanMapping(nullValuePropertyMappingStrategy = NullValuePropertyMappingStrategy.IGNORE)
+    void updateProductFromDto(ProductCreateDto dto, @MappingTarget Product product);
 
     Product toEntity(ProductCreateDto productCreateDto);
