+++ src/main/java/com/telran/store/service/ProductService.java
@@ -1,5 +1,6 @@
 package com.telran.store.service;
 
+import com.telran.store.dto.ProductCreateDto;
 import com.telran.store.entity.Product;
 
 import java.util.List;
@@ -13,4 +14,6 @@ public interface ProductService {
     Product getById(long productId);
 
     void deleteById(long productId);
+
+    Product edit(Long id, ProductCreateDto productCreateDto);
 }
