+++ src/main/resources/application.properties
@@ -1,4 +1,4 @@
-spring.datasource.url=jdbc:postgresql://localhost:5433/onlinegardenshop
+spring.datasource.url=jdbc:postgresql://localhost:5432/onlinegardenshop
 spring.datasource.username=postgres
 spring.datasource.password=postgres
